+++ zenml/steps/base_step.py
@@ -16,7 +16,7 @@ import inspect
 from typing import Dict
 
 from zenml.backends import BaseBackend
-from zenml.core.standards.standard_keys import StepKeys
+from zenml.standards.standard_keys import StepKeys
 from zenml.utils.enums import StepTypes
 from zenml.utils.print_utils import to_pretty_string, PrintStyles
 from zenml.utils.source_utils import resolve_source_path, \
